+++ src/main/java/com/adaming/myapp/dao/reservation/DaoReservation.java
@@ -1,5 +1,6 @@
 package com.adaming.myapp.dao.reservation;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Logger;
 
@@ -8,6 +9,7 @@ import javax.persistence.PersistenceContext;
 import javax.persistence.Query;
 
 import com.adaming.myapp.entities.*;
+import com.adaming.myapp.exception.ParameterException;
 
 public class DaoReservation implements IdaoReservation {
 
@@ -16,11 +18,29 @@ public class DaoReservation implements IdaoReservation {
 	private EntityManager em;
 
 	@Override
-	public Reservation addReservation(Reservation r, Long idC, Long idCh) {
-		Personne client = em.find(Personne.class, idC);
-		r.setPersonne(client);
+	public Reservation addReservation(Reservation r, Long idC, Long idCh) throws ParameterException {
+		Client c = em.find(Client.class, idC);
+		if(c==null)
+		{
+			throw new ParameterException("Le client referee n'existe pas.");
+		}
+		r.setPersonne(c);
 		Chambre ch = em.find(Chambre.class, idCh);
+		if(ch==null)
+		{
+			throw new ParameterException("La chambre referee n'existe pas.");
+		}
 		r.setChambre(ch);
+		for(Reservation rr:ch.getReserv())
+		{
+			if(rr.getDateArrivee().before(r.getDateArrivee()) && rr.getDateSortie().after(r.getDateArrivee())|| 
+					rr.getDateArrivee().before(r.getDateSortie()) && rr.getDateSortie().after(r.getDateSortie())||
+					rr.getDateArrivee().after(r.getDateArrivee()) && rr.getDateSortie().before(r.getDateSortie())||
+					rr.getDateArrivee().before(r.getDateSortie()) && rr.getDateSortie().after(r.getDateArrivee()))
+			{
+				throw new ParameterException("La chambre a deja ete reservee pour cette date");
+			}
+		}
 		em.persist(r);
 		log.info("La reservation " + r.getIdReservation()
 				+ " a bien ete enregistre .");
@@ -29,7 +49,11 @@ public class DaoReservation implements IdaoReservation {
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public List<Reservation> getReservations(Long idC) {
+	public List<Reservation> getReservations(Long idC) throws ParameterException {
+		if(em.find(Client.class, idC)==null)
+		{
+			throw new ParameterException("Le client referee n'existe pas.");
+		}
 		Query req = em.createQuery("from Reservation where idPersonne=:idC");
 		req.setParameter("idC", idC);
 		log.info("Il existe une liste de " + req.getResultList().size()
@@ -38,29 +62,41 @@ public class DaoReservation implements IdaoReservation {
 	}
 
 	@Override
-	public Reservation deleteReservation(Long idR) {
+	public Reservation deleteReservation(Long idR) throws ParameterException {
 		Reservation r = em.find(Reservation.class, idR);
+		if(r==null)
+		{
+			throw new ParameterException("La reservation referee n'existe pas.");
+		}
 		em.remove(r);
 		log.info("La reservation "+r.getIdReservation() +" a bien ete supprime.");
 		return r;
 	}
 
 	@Override
-	public Reservation updateReservation(Reservation r) {
+	public Reservation updateReservation(Reservation r) throws ParameterException {
+		if(r.getIdReservation()==null)
+		{
+			throw new ParameterException("La reservation referee n'existe pas.");
+		}
 		em.merge(r);
 		log.info("La reservation "+r.getIdReservation()+ " a bien ete modifie.");
 		return r;
 	}
 
 	@Override
-	public Double getTotalCostReservation(Long idR) {
+	public Double getTotalCostReservation(Long idR) throws ParameterException {
 		Reservation r = em.find(Reservation.class, idR);
-		Double total = r.getChambre().getPrix();
-		for(Consommation co:r.getConsom())
+		if(r==null)
 		{
-			for(int i=0;i<co.getQuantiteConsommee();i++)
+			throw new ParameterException("La reservation referee n'existe pas.");
+		}
+		Double total = r.getChambre().getPrix();
+		if(!(r.getConsom().equals(null)))
+		{
+			for(Consommation co:r.getConsom())
 			{
-				total=co.getProduit().getPrixProduit()+total;
+				total=(co.getProduit().getPrixProduit()*co.getQuantiteConsommee())+total;
 			}
 		}
 		return total;
@@ -75,11 +111,11 @@ public class DaoReservation implements IdaoReservation {
 		for(Reservation r:tab)
 		{
 			total = r.getChambre().getPrix()+total;
-			for(Consommation co:r.getConsom())
+			if(!(r.getConsom().equals(null)))
 			{
-				for(int i=0;i<co.getQuantiteConsommee();i++)
+				for(Consommation co:r.getConsom())
 				{
-					total=co.getProduit().getPrixProduit()+total;
+					total=(co.getProduit().getPrixProduit()*co.getQuantiteConsommee())+total;
 				}
 			}
 		}
@@ -100,28 +136,70 @@ public class DaoReservation implements IdaoReservation {
 				if(c.getProduit()==p)
 				{
 					p.setQuantiteProduite(c.getProduit().getQuantiteProduite()-c.getQuantiteConsommee());
+					if(p.getQuantiteProduite()<0)
+					{
+						p.setQuantiteProduite(0);
+					}
 				}
 			}
 		}
 
 		return ptab;
 	}
+	
+//	@SuppressWarnings("unchecked")
+//	@Override
+//	public List<Produit> getStock() {
+//		Query req = em.createQuery("p.quantiteProduite-c.quantiteConsommee as stock from Produit p, Consommation c");
+//		List<Produit> ptab = req.getResultList();
+//
+//		return ptab;
+//	}
 
 	@SuppressWarnings("unchecked")
 	@Override
-	public List<Reservation> getReservationParResa(Long idR) {
+	public List<Reservation> getReservationParResa(Long idR) throws ParameterException {
+		if(em.find(Reservation.class, idR)==null)
+		{
+			throw new ParameterException("La reservation referee n'existe pas.");
+		}
 		Query req = em.createQuery("from Reservation where idReservation=:idR");
 		req.setParameter("idR", idR);
 		log.info("La réservation " +idR+ " a bien été récupéré!");
 		return req.getResultList();
 	}
 
+	@SuppressWarnings("unchecked")
 	@Override
-	public List<Reservation> getReservationParChambre(Long idCh) {
+	public List<Reservation> getReservationParChambre(Long idCh) throws ParameterException {
+		if(em.find(Chambre.class, idCh)==null)
+		{
+			throw new ParameterException("La chambre referee n'existe pas.");
+		}
 		Query req = em.createQuery("from Reservation where idChambre=:idCh");
 		req.setParameter("idCh", idCh);
 		log.info("La réservation de la chambre " +idCh+ " a bien été récupéré! et il y a "+req.getResultList().size()+" réservation sur cette chambre!");
 		return req.getResultList();
+	}
+
+	@SuppressWarnings("unchecked")
+	@Override
+	public List<Reservation> getReservationParHotel(Long idH) throws ParameterException {
+		if(em.find(Hotel.class, idH)==null)
+		{
+			throw new ParameterException("L'hotel referee n'existe pas.");
+		}
+		Query req = em.createQuery("from Reservation");
+		List<Reservation> tab = req.getResultList();
+		List<Reservation> tabH = new ArrayList<Reservation>();
+		for(Reservation r:tab)
+		{
+			if(r.getChambre().getHotel().getIdHotel()==idH)
+			{
+				tabH.add(r);
+			}
+		}
+		return tabH;
 	}
 
 }
